requirements.txt
numpy==1.26.4
pandas
pyyaml
scapy
tensorflow==2.15.0
